Refatoração de testes e a criação dos que faltavam

+++ StoreManager/src/StoreManager.Infrastructure/Repositories/FunctionRepository.cs
@@ -53,6 +53,11 @@ namespace StoreManager.Infrastructure.Repositories
         {
             var foundFunction = await context.Functions.FindAsync(id);
 
+            if (foundFunction == null)
+            {
+                return null;
+            }
+
             context.Functions.Remove(foundFunction);
             await context.SaveChangesAsync();
             return foundFunction;
